refactor(onboarding): 步骤组件拆分到 steps/ 各自成文件
OnboardingWizard.tsx 的四个步骤组件（欢迎 / 平台 / LLM / 完成）从单文件
拆到 onboarding/steps/ 各自成文件，容器（485→236 行）只留向导骨架：
步骤指示 + slide 切换 + 底部导航 + finish/onLanguageChange。

容器路径与对外接口（OnboardingWizard / OnboardingResult）不变，纯搬运无
逻辑/界面变更。
